TP_2/requirements.txt
requests
beautifulsoup4
Flask
aiohttp
aiohttp-retry
selenium
webdriver-manager
Pillow
lxml
aiofiles
cairosvg